pyproject.toml
[tool.poetry]
name = "finance-bot"
version = "0.1.0"
description = "Telegram bot for personal finance tracking"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "bot" }
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = [
    "tests"
]
python_files = [
    "test_*.py"
]
asyncio_mode = "auto"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
SQLAlchemy = "^2.0.7"
alembic = "^1.10.2"
magic-filter = "^1.0.9"
aiosqlite = "^0.19.0"
cachetools = "^5.3.3"
pytest-mock = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
mypy = "^1.3.0"
ruff = "^0.0.272"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"